sale4cast_V_0.0
This is a first version of sale4cast. This version was tested on regular Hourly Spaced Data, Irregular minute space data and Outlier Irregular Minute Space Data file. Problem: Visualization of a big data such as a data of 1 year is very difficult. Especially time complexity is very high to produce a forecast. Next Step: Slice a big data by a tail of last 2 months and then work on it for visualization and forecast.