Next step from last commit was not done. Rather plot was made as mobile friendly.
Slice a big data by a tail of last 2 months and then work on it for visualization and forecast is not fixed. Rather the code to generate all of plot was transferred from server.R to it's individual .R plot file. Especially, plot legend and title was not fitted in the mobile screen. Hence the title of a plot was covered by HTML tag. A lot of study was performed on GGplot legend and scale and dynamic UI. Next step: create a hidden input slider to select a month range from the last observation. Slice a big data by a tail of last 2 months and then work on it to provide a forecast.