Add a date slider and repair daily forecast accuracy.
Created a hidden input date slider to select a month range from the last observation. Slice a big data by a tail of last 4 months and then work on it to provide a forecast. There was some mistake in generating daily forecast, especially a train set was truncated where data was cleaned for whole train and test set beforehand. If there is an extreme  value on a test set then it also affect on the interpolation of train set. Therefore a train set was only cleaned by applying tsclean() before forecasting a train set. This forecast data was compared with original test data. Next Step: more test should be performed on minute data by collecting different different minute data set.